+++ src/assets/index.tsx
@@ -1,3 +1,4 @@
-import FilterIcon from './filter.svg?react';
+import FilterIcon from './Icon/filter.svg?react';
+import SearchIcon from './Icon/search.svg?react';
 
-export { FilterIcon };
+export { FilterIcon, SearchIcon };
