--- src/components/layout/searchBar/SearchBar.tsx
@@ -1,5 +1,0 @@
-const SearchBar = () => {
-  return <div>SearchBar</div>;
-};
-
-export default SearchBar;
